fix(client): isolate catalog uniqueness harness from public surface
- Move generation-time uniqueness + trybuild into publish=false
  openlark-capability-unique (single implementation; no openlark-client
  test feature / macro_export)
- Fix CHANGELOG #434 note and ARCHITECTURE draft pointer to capability
  catalog (remove stale registry/catalog.rs narrative)
- Split Client catalog contract tests into catalog_contract_tests.rs
