+++ CHANGELOG.md
@@ -32,9 +32,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   Client/registry 双声明。禁用 feature 时两处均不产生字段或 entry。
 
 - **client：编译期能力目录 tracer（#434 / #423）**：新增 `capability` catalog，
-  以 `bot` 为 tracer 用同一声明同时生成 `Client::bot` 字段与 registry 诊断元数据；
-  启用 `bot` feature 时 `client.bot` 可用且 `registry.has_service("bot")`；禁用时
-  两处均不可用。其余业务域仍走 `declare_client!` + `registry/catalog.rs` 旧路径。
+  以 `bot` 为 tracer 起步；随后 #435/#436 将全部业务域迁入同一目录（见上）。
+  启用对应 feature 时 Client 字段与 `registry.has_service` 一致；禁用时两处均无。
 
 - **core：加深 `Transport::request` 请求执行（#422 / #430–#433）**：
   - 内部收敛为 `request_execution` 深模块（构建 + 认证 + 解码）；删除纯委托
