+++ Cargo.toml
@@ -4,6 +4,8 @@ members = [
     # === 核心基础设施 (Core Infrastructure) ===
     "crates/openlark-core",           # 核心 HTTP 客户端和错误处理
     "crates/openlark-client",         # 高级客户端接口
+    # 内部：capability catalog 生成期唯一性（publish=false，非 SDK 公开面）
+    "crates/openlark-capability-unique",
 
     # === 生产就绪模块 (Production Ready) ===
     # 认证与安全
