+++ crates/openlark-capability-unique/tests/ui.rs
@@ -1,6 +1,4 @@
-//! compile-fail 回归：capability catalog generation-time 唯一性（#423 / #455）
-//!
-//! 需 `--features __compile_fail_harness`（本文件 `required-features` 已声明）。
+//! compile-fail：generation-time catalog uniqueness（#423）
 
 #[test]
 fn capability_catalog_generation_time_uniqueness() {
