+++ crates/openlark-capability-unique/tests/ui/mismatch_capability_name.rs
@@ -1,0 +1,7 @@
+//! name 与 field 不一致必须在生成期失败
+
+fn main() {
+    openlark_capability_unique::assert_capability_catalog_unique! {
+        { field: auth, name: "bot" },
+    }
+}
