+++ crates/openlark-client/Cargo.toml
@@ -17,6 +17,7 @@ workspace = true
 [dependencies]
 # Core dependencies
 openlark-core = { workspace = true }
+openlark-capability-unique = { path = "../openlark-capability-unique" }
 serde = { workspace = true }
 serde_json = { workspace = true }
 tokio = { workspace = true }
@@ -85,18 +86,9 @@ websocket = ["tokio-tungstenite", "futures-util", "lark-websocket-protobuf", "pr
 core-layer = ["communication", "docs", "security"]
 p0-services = ["communication", "docs", "security"]
 
-# 仅 trybuild compile-fail 导出隐藏宏；非产品能力（见 tests/ui.rs required-features）
-__compile_fail_harness = []
-
 [dev-dependencies]
 tokio-test = "0.4"
 wiremock = { workspace = true }
 rstest = { workspace = true }
 tempfile = { workspace = true }
 mockall = { workspace = true }
-trybuild = "1.0"
-
-[[test]]
-name = "ui"
-path = "tests/ui.rs"
-required-features = ["__compile_fail_harness"]
